requirements.txt
langchain
ipykernel
python-dotenv
cohere
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
langchain-cohere
torch==2.3.1
sentence_transformers
langchain_huggingface
chromadb
langchain_chroma
langchain-google-vertexai
langchain-ollama
faiss-cpu
streamlit
langchain_groq
uvicorn
fastapi
langserve
transformers
python-multipart
llama-index
langchain-classic
unstructured[pdf]

